refactor NewVirtualService
