+++ internal/controller/watcher_controller.go
@@ -50,6 +50,7 @@ type WatcherReconciler struct {
 	client.Client
 	record.EventRecorder
 	IstioClient        *istio.Client
+	IstioService       *istio.Service
 	WatcherVSNamespace string
 	RestConfig         *rest.Config
 	Scheme             *machineryruntime.Scheme
@@ -144,7 +145,8 @@ func (r *WatcherReconciler) handleDeletingState(ctx context.Context, watcherCR *
 func (r *WatcherReconciler) handleProcessingState(ctx context.Context,
 	watcherCR *v1beta2.Watcher,
 ) (ctrl.Result, error) {
-	virtualSvc, err := r.IstioClient.NewVirtualService(ctx, watcherCR, r.WatcherVSNamespace)
+
+	virtualSvc, err := r.IstioService.NewVirtualServiceForWatcher(ctx, watcherCR, r.WatcherVSNamespace)
 	if err != nil {
 		return r.updateWatcherState(ctx, watcherCR, shared.StateError, err)
 	}
